requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
boto3==1.39.8
botocore==1.39.8
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.116.1
fortls==3.2.2
greenlet==3.2.3
h11==0.16.0
idna==3.10
jmespath==1.0.1
json5==0.12.0
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.13.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
wheel==0.45.1
